avidflow-back/requirements.txt
aio-pika==9.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.34.69
botocore==1.34.162
build==1.3.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.20.0
croniter==6.0.0
cryptography==42.0.5
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
ephem==4.2
fastapi==0.115.12
fastapi-pagination==0.13.1
fastapi-storages @ git+https://github.com/aminalaee/fastapi-storages.git@edc624fb647e128e82669f19b195e670a4714dfd
fastavro==1.12.1
feedparser==6.0.11
filelock==3.20.0
frozenlist==1.6.0
fsspec==2025.10.0
gevent==24.11.1
googleapis-common-protos==1.72.0
greenlet==3.2.1
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.1.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.27
langchain-cohere==0.4.6
langchain-community==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langfuse==3.10.1
langsmith==0.4.42
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.7.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
pip-tools==7.5.2
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.33.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.11
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
redis==5.0.3
regex==2025.11.3
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.10.4
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlalchemy_celery_beat==0.8.4
starlette==0.46.2
starlette-admin==0.15.1
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.0
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
zstandard==0.25.0
